Preview 6 refactor
